refactor(frontend): move BTC address utils to lib
